Update response for track data
